+++ DefiLecture-JC/src/java/com/defilecture/controleur/AfficherPageModificationDefiAction.java
@@ -35,7 +35,7 @@ public class AfficherPageModificationDefiAction extends Action {
           Defi d = new DefiDAO(cnx).read(request.getParameter("id"));
 
           // seul celui qui a ajouté la lecture peut la modifier
-          if (d != null && d.getIdCompte() == (int) session.getAttribute("currentId")) {
+          if (d != null && d.getIdCompte() == ((Integer) session.getAttribute("currentId")).intValue()) {
             request.setAttribute("vue", "pageModificationDefi.jsp");
           }
         }
