+++ DefiLecture-JC/src/java/com/defilecture/controleur/EffectuerSuppressionBDAction.java
@@ -41,7 +41,7 @@ public class EffectuerSuppressionBDAction extends Action implements RequirePRGAc
 
   @Override
   public String execute() {
-    if ((userIsConnected()) && (int) session.getAttribute("role") > Compte.MODERATEUR) {
+    if ((userIsConnected()) && ((Integer) session.getAttribute("role")).intValue() > Compte.MODERATEUR) {
       try {
         Connexion.reinit();
         Connection cnx =
@@ -55,7 +55,7 @@ public class EffectuerSuppressionBDAction extends Action implements RequirePRGAc
         cSiteDao = new ConfigSiteDAO(cnx);
 
         // Vérification du mot de passe de l'usager administrateur
-        Compte compteAdmin = daoCompte.read((int) session.getAttribute("currentId"));
+        Compte compteAdmin = daoCompte.read(((Integer) session.getAttribute("currentId")).intValue());
         Compte verif =
             daoCompte.findByIdentifiantMotPasse(
                 compteAdmin.getPseudonyme(), request.getParameter("passwordConf"));
